pyproject.toml
[project]
name = "craft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beanie>=1.29.0",
    "fastapi>=0.115.12",
    "logfire[fastapi,pymongo]>=3.14.0",
    "passlib>=1.7.4",
    "pydantic>=2.11.3",
    "pydantic-ai>=0.1.3",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
]

[project.scripts]
craft = "craft.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "uvicorn>=0.34.2",
]
lint = [
    "ruff>=0.11.6",
]
